refactor: LabelChip 통합 및 범용 컴포넌트로 구조 개선
- 중복되는 ColoredLabelChip 삭제 및 LabelChip으로 기능 통합
- LabelChip이 텍스트 외에 다양한 Content를 수용할 수 있도록 변경
- 배경색 및 패딩 설정을 위한 파라미터 추가로 유연성 확보
