+++ app/src/main/java/com/boostcamp/and03/ui/screen/bookdetail/component/CharacterCard.kt
@@ -101,7 +101,15 @@ fun CharacterCard(
 
                     Spacer(modifier = Modifier.height(And03Spacing.SPACE_XS))
 
-                    LabelChip(text = role)
+                    LabelChip(
+                        content = {
+                            Text(
+                                text = role,
+                                style = MaterialTheme.typography.labelSmall,
+                                color = And03Theme.colors.onSecondaryContainer
+                            )
+                        }
+                    )
                 }
 
                 DropdownMenuContainer(
